+++ akka-tools-jdbc-journal/src/main/scala/no/nextgentel/oss/akkatools/persistence/jdbcjournal/StorageRepo.scala
@@ -61,7 +61,7 @@ class StorageRepoImpl(sql2o: Sql2o, schemaName: Option[String], _errorHandler:Jd
   def loadJournalEntries(persistenceId: PersistenceId, fromSequenceNr: Long, toSequenceNr: Long, max: Long): List[JournalEntryDto] = {
     val sequenceNrColumnName = persistenceId match {
       case p:PersistenceIdSingle    => "sequenceNr"
-      case p:PersistenceIdTagOnly => "journalIndex"
+      case p:PersistenceIdTagsOnly => "journalIndex"
     }
 
     val preSql = s"select * from (select typePath, id, $sequenceNrColumnName, persistentRepr, updated from ${schemaPrefix}t_journal where "
@@ -78,9 +78,12 @@ class StorageRepoImpl(sql2o: Sql2o, schemaName: Option[String], _errorHandler:Jd
               .addParameter("typePath", p.tag)
               .addParameter("id", p.uniqueId)
 
-          case p:PersistenceIdTagOnly =>
+          case p:PersistenceIdSingleTagOnly =>
             conn.createQuery(preSql + " typePath = :typePath " + postSql)
               .addParameter("typePath", p.tag)
+
+          case p:PersistenceIdMultipleTags =>
+            conn.createQuery(preSql + " typePath in (" + p.tags.map( s => "'" + s + "'" ).mkString(",") + ") " + postSql)
         }
         query.addParameter("fromSequenceNr", fromSequenceNr).addParameter("toSequenceNr", toSequenceNr).addParameter("max", max)
 
@@ -157,7 +160,7 @@ class StorageRepoImpl(sql2o: Sql2o, schemaName: Option[String], _errorHandler:Jd
 
     val sequenceNrColumnName = persistenceId match {
       case p:PersistenceIdSingle    => "sequenceNr"
-      case p:PersistenceIdTagOnly => "journalIndex"
+      case p:PersistenceIdTagsOnly  => "journalIndex"
     }
 
     val preSql = s"select max($sequenceNrColumnName) from ${schemaPrefix}t_journal where "
@@ -171,9 +174,13 @@ class StorageRepoImpl(sql2o: Sql2o, schemaName: Option[String], _errorHandler:Jd
           c.createQuery( preSql + " typePath = :typePath and id = :id " + postSql)
             .addParameter("typePath", p.tag)
             .addParameter("id", p.uniqueId)
-        case p:PersistenceIdTagOnly =>
+
+        case p:PersistenceIdSingleTagOnly =>
           c.createQuery( preSql + " typePath = :typePath " + postSql)
             .addParameter("typePath", p.tag)
+
+        case p:PersistenceIdMultipleTags =>
+          c.createQuery( preSql + " typePath in (" + p.tags.map( s => "'" + s + "'").mkString(",") + ") " + postSql)
       }
 
       query.addParameter("fromSequenceNr", fromSequenceNr)